almost 100% finished

+++ controller/deleteSharedFile.php
@@ -2,18 +2,23 @@
 
 require("../utils/config.php");
 require("../model/Documents.php");
+require("../model/Permissions.php");
 
 
 $fileModel = new Documents();
-
+$permiss = new Permissions();
 
 $fileId = $_GET['id'] ?? false;
 
+session_start();
 
-$fileModel->delete($fileId);
+$user_id = $_SESSION['user'];
 
-header('location: http://localhost/pw/trabalho-programa-o-web/controller/sharedFiles.php?error=1');
+$permitted = $permiss->auth_shared(intval($user_id['id']), intval($fileId));
 
-
-
-?>
+if ($permitted === true) {
+  $fileModel->delete($fileId);
+  header('location: http://localhost/pw/trabalho-programa-o-web/controller/sharedFiles.php?error=1');
+} else {
+  header('location: http://localhost/pw/trabalho-programa-o-web/controller/sharedFiles.php?error=2');
+}
